Once the correct answer is revealed, it is always revealed.
Each time a feedback popover is closed and the opened again, the correct
answer "Reveal" button is back again.  So this makes it so that once the
correct answer has been revealed once, each subsequent time the popover
is opened it is automatically revealed without needing to click the
reveal button again.  Note that a page reload will reset this behavior.