Make the MathQuill editor toolbar appear on the right even for right-to-left languages.
Currenlty if the course language is set to he-IL and the
`$perProblemLangAndDirSettingMode` is set to `auto:en:ltr`, then the
computations for the toolbar position make the toolbar on the right but
way to wide.  This just makes the toolbar appear the same as for
left-to-right languages.  That is not strictly correct, but since the
toolbar is not currently translated, I am not sure much better can be
done at this point.

This addresses issue #988.