Add `strict` to modules shared to the PG safe zone in pg_config.dist.yml.
Versions of Perl prior to 5.34 need this.  Otherwise warnings are issue
by the `BEGIN { strict->import }` approach now used in place of the old
`BEGIN { be_strict() }` approach.

Note that this is not used by webwork2, but the standalone renderer will
need this (if anyone runs that with an older perl version).  So the
standalone renderer will also need to update its distributed
pg_config.yml file.

Also change `Main` to `main` in the translator.  I can't resist fixing
this any longer.  The proper namespace is `main`.  This is also used
inconsistently in `Translator.pm`.  Other than these two instances
`main` is used everywhere else.