Fix the feedback class elements for checkbox or radio part answers of a RadioMultiAnswer.
This was observed while looking at issue #1007 (but is not that issue).

When a RadioMultiAnswer has checkbox or radio button parts, then the
feedback class is only being added to the first checkbox or radio
button.  This updates the `parserRadioMultiAnswer.pl` `feedback_options`
method to look for these, and correctly mark those that are checked (or
all of them if none are checked).