Update JSXGraph to version 1.7.0.
This fixes the issue observed in #1014.  Although there is another issue
with a console.log statement that seems to have accidentally left in the
release.  This console.log message occurs anytime the coordinate hints
change.  So for problems that use the `showCoordinateHints => 0` option,
this message doesn't appear!  This has been fixed in their main branch,
so hopefully they release a patch fix soon that includes this fix.
Hopefully before our next release.