Fix Value::Error calls in a checker for a RadioMultiAnswer object.
This fixes the issue observed in
https://webwork.maa.org/moodle/mod/forum/discuss.php?d=8489.  See the
problem posted there for a problem to test this with. There were two
problems really that caused the issue.  First `$self->{ans}` is a doubly
nested array.  So `$self->{ans}[0]` won't be an answer evaluator, but
rather an array of answer evaluators.  That should have been
`$self->{ans}[0][0]`.  However, that still won't work in this case,
because for the problem in that forum post there are no answers in the
first part.  Thus the `$self->{ans}[0]` is the emtpy array, and so
`$self->{ans}[0][0]` is undefined, and still not an answer evaluator.
So another approach was taken, and the context error is just added via
the RadioMultiAnswer `appendMessage` instead.

This also adds the context swapping added for MultiAnswer objects in
pull request #632.  The same issues that led to that pull request can
occur for a RadioMultiAnswer.

Furthermore, this fixes an issue in the changes from #632.  That is that
the `$ans` used on line 307 of parserMultiAnswer.pl is a variable that
is not declared. There is no error for this because `strict` is not
(yet) in effect for this macro.