Accessibility tweak for the MathQuill toolbar.
This adds the `aria-hidden` attribute to the span containing the static
MathQuill element in the toolbar buttons, and changes the
`data-bs-title` attribute of the button itself to a `title`. I don't
know if the latter change makes much difference since bootstrap will
move the title from the button to the tooltip in any case.  But adding
`aria-hidden` will prevent the contents of the span from being read by
screen readers.  The title of the tooltip or button should be read
instead.

Maybe this will work to address https://github.com/openwebwork/mathquill/issues/24?