Switch to using `[< ... >]` instead of `[^ ... ^]` for tags.
Change the general syntax to

```
[< content >]{
    html => [ 'tag_name', class => 'my-class', ... ],
    tex  => [ 'tex begin', 'tex end' ],
    ptx  => [ 'tag name', { attribute => 1, ... }, 'separator' ]
}
```

or

`[< content >]{['tag_name', class => 'my-class', ...]}{['tex begin', 'tex end']}{['tag name', { attribute => 1, ... }, 'separator']}`

The `html`, `tex`, and `ptx` options can also just be strings.  For
`html` and `ptx` that string will be the tag name.  If the `tex`
argument is a string then the contents will be wrapped in a TeX
environment by that name.

Note that for `html` when the argument is given as an array, the tag
name can also be omitted.  In that case a `div` tag will be used.  So
you can do `[< content >]{[ class => 'my-class' ]}`.

A link can be created with this as well.  For example,
`[<Google>]{[ 'a', href => 'https://www.google.com' ]}`.

I would be willing to switch the default to being an `a` tag.  So then
to get a `div` you would need `[< content >]{[ 'div', class => 'my-class' ]}`,
but to obtain a link you could do
`[<Google>]{[ href => 'https://www.google.com' ]}`.  That would make
the `[< ... >]` block more like the originaly intended link. I left the
default as a `div` because that usage is probably nicer for what I would
use this for (I never put links in problems).

For now the only allowed HTML tags are `div`, `span`, and `a`.  What
other tags should be allowed?

Note that for `PTX` display mode this uses the `NiceTables::tag` method.
I am not sure if that was a good idea, but it was an idea.
@Alex-Jordan: Please advise.