Massive clean up of PGalias.pm and PGresource.pm.
Rewrite the PGalias.pm and PGresource.pm files so that the implemenation
makes much more sense and doesn't have the massive kludge that they
were.

The behavior is the same as it was.  The structure of the data stored in
a PGresource object is different (all of the things that were only used
internally by PGalias are no longer there), and a minor change to
webwork2 is needed to accomodate for that cchange.

There is one change of note.  This switches PGalias.pm from using the
"gif2png" external program to using the "convert" external program to
convert gif files to png files.  Note that the "giv2png" external
program is "/usr/bin/giftopnm" by default.  Of course "convert" is
imagemagick's convert utility. This works just as well (if not better).
Note that an animated gif is converted to the first frame (same as
before).

Note that "csv" files are also added to the list of allowed auxiliary
resources. At this point this is not used, but the rserve code needs to
be updated to use PGalias.  So that is what this is for. I see no reason
that a csv file should not be allowed in any case.  An author may want
to provide a static csv data file with a problem.

The copy of the `findMacroFile` that was in PGalias.pm has been deleted.
The PG.pl `findMacroFile` calls the method in PGloadfiles.pm instead.
Note that the copy in PGalias.pm was not entirely functional to begin
with.  Also note that the version in PG.pl is only called by the
deprecated answerDiscussion.pl macro.  The only other places where this
method is called is by PGloadfiles.pm internally (for actually loading
macro files).

The `findAppletCodebase` stub in PGalias.pm was also deleted.  The only
place that is called is by the `findAppletCodebase` method in PG.pl.  So
that method is now the stub directly.