Cleanup of PGloadfiles.pl.
This is just cleanup other than the removal of the `sourceAlias` method.
That method is not used, and returns a completely dysfunctional link.
It has nothing to do with the `source.pl` macro or how it works, and is
not needed for the functionality of that macro (which does still work if
proxying via apache2).