Perform a bit of html validation for PGML tag blocks that are spans.
A span tag block is not allowed to contain any PGML blocks that have the
type indent, align, par, list, bullet, answer, heading, rule, code, pre,
verbatim, table, or tag.  If a span tag block is detected with any of
those things in it, a warning is issued and the contents rendered
directly without the span tag.  This helps to prevent problem authors
from doing the wrong thing.

It is valid for a span to contain a span, but if a span tag block
contains another tag bloc, the outer span tag block does not have the
information to determine what the inner tag block is (span or div).
Since I don't think it is particularly useful to have a span within a
span, I just blocked any tag block in a span.