Change knowls to being buttons that open modal dialogs.
This definitively solves the problem of placement of the knowl content.
The old method of trying to find a place to put the content in an
accordion is entirely unreliable, and depends on the problem author not
doing bad things.

Technically a knowl link is not a link at all anymore.  In fact it is
now a button.  Although, it really wasn't a link before either.  It was
an anchor with a href to the root fragment, and the default action for
that was disabled by JavaScript.

There is a new option that can be passed to the `knowlLink` method.
That is a `title`.  This `title` will be used for the title of the modal
dialog if provided.  Otherwise first required argument of the
`knowlLink` (the `$display_text`) will be used for the modal dialog
title.