Fix a couple of minor issues with feedback.
First, the `previewAnswer` and `feedbackLine` methods should not be
nested subs.  They should be closures named by a variable instead.

Second, the `aria-describedby` attribute added to the inputs needs to be
removed since its value is for an element that is not in the DOM until
the feedback popover is opened.