Update dependencies again.
Most importantly a new tag has been created in the MathQuill repository
for WeBWorK/PG 2.19, and the package.json file version of mathquill now
points to that. Note that the tag will need to be updated if/when the
pending pull requests to that repository are merged, but the change to
the package.json file will still work.

Other updates are:

* plotly.js-dist-min: 2.27.0 to 2.32.0
* sortablejs: 1.15.0 to 1.15.2
* autoprefixer: 10.4.16 to 10.4.19
* chokidar: 3.5.3 to 3.6.0
* cssnano: 6.0.1 to 6.1.2
* postcss: 8.4.31 to 8.4.38
* prettier: 3.1.1 to 3.2.5
* sass: 1.69.5 to 1.75.0
* terser: 5.24.0 to 5.30.0

Note that jsxgraph is NOT updated to the latest version.  The latest
version is 1.8.0, and we are using 1.7.0.  Although 1.7.0 has a console
log issue that is annoying, that is better than the issue with the grid
that I see in 1.8.0 that makes the grid thick and obtrusive in certain
situations.  I reported the issue, and it has been fixed in their
develop branch, but that is not in a release yet.