Update the package-lock.json to point to the latest MathQuill commit.
Pull request https://github.com/openwebwork/mathquill/pull/28 was merged
into MathQuill.  That commit is now tagged for WeBWorK-2.19.  So if you
run `npm install` it will install that commit.  However, if you run `npm
ci` it will get the commit before it because that is what is in the
package-lock.json file.