Add an en-GB localization file.
This isn't exactly needed, but is added for consistency.  As for
webwork2, translation isn't really needed in the `.po` file.  However,
if someone wants to do so it could be done.

Note that the `en.po` file was updated as well.  This file hasn't been
updated for a while since it really doesn't need to be.