Use `@openwebwork/mathquill` from npm instead of directly using github.
A beta version of an `@openwebwork/mathquill` package has been published
on npm.  This switches to using that instead of from github.

The github repository will also be updated to not include the
distribution build anymore.  This fixes the mess of always needing to
reubild the `dist` files for every change to the mathquill code.  That
is a real pain when different pull requests are made, particularly
unrelated pull requests, as they have to be put in order with sequential
builds.