Improvements to the debugging `show` method of PGML.
This was something that was useful for me while I was working on writing
the CodeMirror 6 PGML parser.  It just formats the things in the PGML
stack that are hashes in a nicer way, actually showing the contents
instead of something like `HASH(0x5d3919403d78)` as it previously did.

This does not affect normal PGML usage in problems at all.