Move feedback button for one by one matrix out of matrix.
A one by one matrix does not have a last answer rule with the MaTrIx
prefix or the row and column indices for a suffix.  In this case, the
last answer name is of course also the first answer name.  As such the
data attribute added was not naming the correct answer rule.

This fixes issue #1129.