Update WeBWorK::PG::Tidy to handle content after ENDDOCUMENT.
The content is removed but saved before perltidy is called.  Then put
back after perltidy is done.  So its contents will not change.  That is
with the exception of possibly adding `();` if that is not there.  Note
that if you have something like `ENDDOCUMENT('invalid arg');` the
`'invalid arg'` will be removed.