Change the order of the handling of an odd number backslashes in post processing.
Instead of placing the `~~` first, place it last.  So if the original
problem has `\\~~`, that becomes `\\\\\` in pre processing.  Then it
becomes `\\~~` again in post processing.  However, if the original
problem has `~~\\`, then that also becomes `\\\\\` in pre processing,
but becomes `\\~~` in post processing.

This is for correctness of escaping when interpreting `~~` as the escape
sequence for PG.  After translation both `\~~` and `~~\` becomes `\\\`,
so the effect is identical.  However, what that means after translation
is that the first backslash escapes the second, and the third backslash
escapce whatever is after it.  So `\~~` is more accurate in meaning when
thinking of `~~` as the escape sequence because the first `\` is now
`\\`, and the `~~` escapes what is next the same as the last `\` does
after translation.