Add default checker to parserMultiAnswer.
This adds a simple default checker to parserMultiAnswer that is used
if no checker is provided. The checker checks if each answer part
is correct using the overloaded `==` operator. It can also be configured
to either return which parts are correct, allowing for partial credit,
or only return 0 or 1 depending on if all answer parts are correct.
The `partialCredit` option can be used to control which behavior
to use, and defaults to the value of `$showPartialCorrectAnswers`.

This is just for convince when using MultiAnswer with a simple
checker.