Fix issue with variable TeX output with var = value inequality.
In contextInequalities, the $context variable could be undefined,
making it so the TeX formatting of a variable in inequalities
of the form Inequality("var = value") was not used. This defines
the variable $context = $self->context in the string and TeX methods
of Inequalities::Set to ensure any variable definitions are used
in the string and TeX output.