Switch to using the latest Perl::Tidy (version 20240903).
This commit just makes the changes in the scripts, but does not yet run
perltidy with the new setup.  That will be done in the next commit.

A new `--valign-signed-numbers` option has been added in the version of
Perl::Tidy and is the default.  That option has been disabled to reduce
the number of code changes.

All of the other changes to the code made by the updated version are
definite improvements.

Note that the `bin/dev_scripts/run-perltidy.pl` script now insists on
this specific version of perltidy instead of this version or newer.
This is because every time a new version of perltidy comes out, it
changes things.  So developers really need to be using the same version
or it will not be consistent with the github workflow result.

Finally, the `-vxl='q'` option to disable alignment of qw quotes has
been removed from the `bin/perltidy-pg.rc` file.  The upside of this is
that those using the `WeBWorK::PG::Tidy` module (either via the
`bin/perldity-pg.pl` script or via webwork2) can use any version of
Perl::Tidy, even those from before the 20220613 version.