Add a `parserLogb.pl` macro for logarithms with base b.
This is similar to the `parserRoot.pl` macro.  To enable the usage of
the `logb` function in the current context call `Parser::Logb->Enable`.

You can call `$n = logb(3, 5)` to get the value of the logarithm with
base 3 evaluated at 5, or `$ans = Compute("logb(3, 5)")` to  use as an
answer.

You can also call `Parser::Logb->EnableComplex` to allow negative bases
and evaluation at negative numbers. Note that attempting to use a
negative base or evaluate at a negative number would otherwise produce
an error message.

An answer in this context will automatically have the MathQuill option
`logsChangeBase => 0` set.  To facilitate this without the need to have
numerous special cases for contexts in the `ENDDOCUMENT` method of
`PG.pl`, there is a new `mathQuillOpts` context flag.  The default value
for this flag is a reference to a hash with no keys.  Any context can
set keys in this hash, and those will be tranferred to the MathQuill
options for any answer rule in the context.  This is now used by the
the `parserRoot.pl` macro to set `rootsAreExponents => 0`, and by the
new `parserLogb.pl` macro to set `logsChangeBase => 0`.

To facilitate students entering a logarithm with a base, if the
`logsChangeBase` option is 0 (or not set), then a subscript button is
added to the MathQuill toolbar.